requirements.txt
Django~=3.1.2
django-filter~=2.4.0
django-crispy-forms
django-mathfilters